requirements.txt
# A-Red-Team-Havoc Requirements
# NIA Red Team Toolkit

# Core dependencies
pyyaml>=6.0
jinja2>=3.1.0
aiohttp>=3.9.0

# Dashboard
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0  # For pretty CLI output
